changing ulong for uint

+++ src/main/kotlin/com/menta_tickets/challenge/responses/ChallengeResponse.kt
@@ -2,4 +2,4 @@ package com.menta_tickets.challenge.responses
 
 import javax.validation.constraints.PositiveOrZero
 
-data class ChallengeResponse(@field:PositiveOrZero val result: ULong)
+data class ChallengeResponse(@field:PositiveOrZero val result: UInt)
